pyproject.toml
[tool.poetry]
name = "my-recipes-shopping-list"
version = "0.1.0"
description = "Recipes & Shopping List Management API"
authors = ["Antonella Barbero <antobarbero94@gmail.com>"]
readme = "README.md"
repository = "https://github.com/antobarbero/my-recipes-shopping-list"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi= "^0.111"
pymongo= "3.11"
pydantic-settings="2.4.0"
python-dotenv="1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^3.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
